Options Menu work
Made the beginnings of the Options Menu; it looks like it will take a
lot of code to get things working. Nothing on the menu is permanent yet.
Also added some background music to the menu screen (holds copywrite  so
I won't be keeping it in very long). Learned a lot about how stupid
Unity is with referencing code.